requirements.txt
chromadb==1.5.8
ollama==0.6.1
pandas==3.0.2
pydantic_settings==2.13.1
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz
streamlit==1.56.0
rouge_score
bert_score
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
